+++ simple-betting-api/src/main/kotlin/com/betting/simplebettingapi/dto/AccountDto.kt
@@ -5,9 +5,16 @@ import javax.validation.constraints.NotNull
 import javax.validation.constraints.Size
 
 class AccountDto(
-    val id: Int = -1,
-    @field:Size(min = 3, max = 10) val username: String,
-    @field:Size(min = 3, max = 10) val name: String,
-    @field:Size(min = 3, max = 10) val surname: String,
+    val id: Long = -1,
+
+    @field:Size(min = 3, max = 10)
+    val username: String,
+
+    @field:Size(min = 3, max = 10)
+    val name: String,
+
+    @field:Size(min = 3, max = 10)
+    val surname: String,
+
     val wallet: WalletDto?
 )
