+++ simple-betting-api/src/main/kotlin/com/betting/simplebettingapi/repository/AccountRepository.kt
@@ -3,5 +3,5 @@ package com.betting.simplebettingapi.repository
 import com.betting.simplebettingapi.model.AccountModel
 import org.springframework.data.repository.CrudRepository
 
-interface AccountRepository : CrudRepository<AccountModel, Int> {
+interface AccountRepository : CrudRepository<AccountModel, Long> {
 }
