+++ simple-betting-api/src/main/kotlin/com/betting/simplebettingapi/repository/BetRepository.kt
@@ -4,8 +4,8 @@ import com.betting.simplebettingapi.model.BetModel
 import com.betting.simplebettingapi.model.RollModel
 import org.springframework.data.repository.CrudRepository
 
-interface BetRepository : CrudRepository<BetModel, Int>{
-    fun findAllByAccountIdOrderByPlacedDtDesc(accountId: Int): List<BetModel>
+interface BetRepository : CrudRepository<BetModel, Long>{
+    fun findAllByAccountIdOrderByPlacedDtDesc(accountId: Long): List<BetModel>
 
     fun findAllByRoll(roll: RollModel): List<BetModel>
 }
